feat: [PAYMCLOUD-35] Add skip_metric_validation to alert criteria and variables (#370)
* Add skip_metric_validation to alert criteria and variables

Introduce the 'skip_metric_validation' option in monitoring alerts criteria and variables. This change provides the ability to optionally disable metric validation, with 'false' set as the default value.

* Fix README formatting for pre tags.

Replaced slash-star lines with slash-fat-arrow in preformatted tags to improve readability and consistency. These changes impact the descriptions of several input variables in the README.