feat: [PAYMCLOUD-198] Emit KSM V1 metrics on opencost module (#399)
* Enable granular control for kube-state-metrics in OpenCost.

Added configuration to disable KSM V1 metrics while enabling KSM V1-only metrics for greater flexibility. This ensures compatibility with specific monitoring setups and cleaner metric management.

* Enable opencost metrics serviceMonitor and update default project.

The change updates the Helm chart value for enabling the serviceMonitor by using the correct key. Additionally, the default project variable is modified from "cstar" to "pagopa" to reflect the updated project naming convention.

* Update Helm chart repository and version for OpenCost

Switch the repository to the official OpenCost Helm chart and upgrade the version to 1.43.0. This change aligns with the latest updates and ensures compatibility with the OpenCost chart source.

* Update Helm chart reference for OpenCost deployment

Replaced the Prometheus OpenCost Exporter chart with the OpenCost chart in the Kubernetes configuration. This ensures alignment with the updated repository and the correct Helm chart for the deployment. Chart version remains configurable as needed.

* Set defaultClusterId for OpenCost in Helm values

Added the `opencost.defaultClusterId` parameter to Helm chart values, assigning it the AKS cluster name. This ensures proper identification of the cluster within OpenCost for accurate cost reporting.

* Disable OpenCost UI and adjust metric emission settings

Updated the configuration to disable the OpenCost UI and set `opencost.metrics.kubeStateMetrics.emitKsmV1MetricsOnly` to false. These changes aim to optimize performance and align with the updated usage requirements.

* Fix typo in OpenCost metric configuration name

Removed an unintended trailing tab in the "emitKsmV1MetricsOnly" name. This ensures the configuration is correctly parsed and functions as expected.

* Downgrade OpenCost Helm chart version to 1.40.0

Revert the OpenCost Helm chart version from 1.43.0 to 1.40.0. This change addresses compatibility issues or potential regressions introduced in the newer version.

* Upgrade OpenCost Helm chart to version 1.43.0

Updated the Helm chart version from 1.40.0 to 1.43.0 to include the latest features, improvements, and bug fixes provided by OpenCost. This ensures the deployment stays up-to-date and aligns with best practices.

* Enable KSM v1 metrics in OpenCost configuration.

Updated the kubeStateMetrics setting to emit KSM v1 metrics by setting the value to "true". This ensures compatibility with systems relying on v1 metric outputs.

* Set Helm chart version for Opencost via variable

Replaced hardcoded Helm chart version with a configurable variable `opencost_helm_chart_version`. This improves flexibility and allows easier updates to the Helm chart version without modifying the main Terraform code. Added a default value for the variable in `99_variables.tf`.